+++ include/wrapper/SERVICES.h
@@ -1,8 +1,6 @@
 #ifndef SWAG_SCANNER_SERVICES_H
 #define SWAG_SCANNER_SERVICES_H
 
-#define TIMER_PAUSE_INTERVAL 10.0
-#define TIMER_SCAN_INTERVAL 5.0
 #define SWAG_SCANNER_NAME @"SwagScanner"
 #define UART_SERVICE_UUID @"5ffba521-2363-41da-92f5-46adc56b2d37"
 #define ROTATE_TABLE_CHAR_UUID @"5ffba522-2363-41da-92f5-46adc56b2d37"
